refactor: move summariser from circuits.py to serialization.py
